TagPOMDPProblem.jl Documentation

Tag POMDP problem using POMDPs.jl. Original problem was presented in Pineau, Joelle et al. “Point-based value iteration: An anytime algorithm for POMDPs.” IJCAI (2003) (online here).

The goal of the agent is to tag the opponent by performing the tag action while in the same square as the opponent. The agent can move in the four cardinal directions or perform the tag action. The movement of the agent is deterministic based on its selected action. A reward of step_penalty is imposed for each motion action and the tag action results in a tag_reward for a successful tag and tag_penalty otherwise. The agent’s position is fully observable but the opponent’s position is unobserved unless both actors are in the same cell. The opponent moves stochastically according to a fixed policy away from the agent. The opponent moves away from the agent move_away_probability of the time and stays in the same cell otherwise. The implementation of the opponent’s movement policy varies slightly from the original paper allowing more movement away from the agent, thus making the scenario slightly more challenging. This implementation redistributes the probabilities of actions that result in hitting a wall to other actions that result in moving away.

TagPOMDPProblem.TagPOMDP — Method

TagPOMDP(; kwargs...)

Returns a TagPOMDP <: POMDP{TagState, Int, Int}. Default values are from the original paper:

Pineau, Joelle et al. “Point-based value iteration: An anytime algorithm for POMDPs.” IJCAI (2003).

The main difference in this implementation is the use of only 1 terminal state and an opponent transition function that aims to keep the probability of moving away to the specified value if there is a valid action (versus allowing the action and thus increasing the probability of remaining in place).

Keywords

tag_grid::TagGrid: Grid details, default = TagGrid()

tag_reward::Float64: Reward for the agent tagging the opponent, default = +10.0

tag_penalty::Float64: Reward for the agent using the tag action and not being in the same grid cell as the opponent, default = -10.0

step_penalty::Float64: Reward for each movement action, default = -1.0

discount_factor::Float64: Discount factor, default = 0.95

move_away_probability::Float64: Probability associated with the opponent srategy. This probability is the chance it moves away, default = 0.8

TagPOMDPProblem.TagGrid — Method

TagGrid(; kwargs...)

Creates a TagGrid struct to contain the grid details for the Tag POMDP. The grid is composed of two parts, a bottome grid and a top grid. The bottom grid x size must be greater than or equal to the top grid x size.

Keywords

bottom_grid::Tuple{Int, Int}: Bottom grid size, default (10,2)

top_grid::Tuple{Int, Int}: Top Grid size, default (3,3)

top_grid_x_attach_pt::Int: Where the bottom left of the top grid attaches to the bottom grid. Only the x component. default 6

TagPOMDPProblem.TagState — Type

TagState

Represents the state in a TagPOMDP problem.

Fields

r_pos::Tuple{Int, Int}: position of the robot

t_pos::Tuple{Int, Int}: position of the target

tagged::Bool: state of target being tagged or not

TagPOMDPProblem.TagPOMDP — Type

TagPOMDP <: POMDP{TagState, Int, Int}

POMDP type for the Tag POMDP.

Fields

tag_grid::TagGrid:

tag_reward::Float64:

tag_penalty::Float64:

step_penalty::Float64:

terminal_state::TagState:

discount_factor::Float64:

move_away_probability::Float64:

TagPOMDPProblem.TagGrid — Type

TagState

Grid details for the Tag POMDP.

Fields

bottom_grid::Tuple{Int,Int}: Bottom grid size

top_grid::Tuple{Int,Int}: Top grid size

top_grid_attach_pt::Tuple{Int,Int}: Where the bottom left of the top grid attaches to the bottom grid

bg_cart_indices::CartesianIndices: Bottom grid cartesian indices

tg_cart_indices::CartesianIndices: Top grid cartesian indices

combined_grid_lin_indices::LinearIndices: Full combined grid linear indicies

full_grid_lin_indices::LinearIndices: Full grid linear indices

full_grid_cart_indices::CartesianIndices: Full grid cartesian indices

exclude_zones::Vector{Vector{Tuple{Int, Int}}}: Vector of zones that are prohibited